MAINT: Silence pooch import dep warning python3.11
Add scipy.datasets files to allowed warning filters
Use warnings.filterwarnings without kw action

Note that when using `warnings.filterwarnings`, a special test needs to
be made aware of the addition of the filtered warning. This test is
the `test_warning_calls_filters` in `test_warnings.py`

It should also be noted that the `test_warning_calls_filters` uses the fixture
`warning_calls` which in turn uses ast to find lines with the use of filtered warnings.
The implementation here is not actually complete since it doesn't account for
`warnings.filterwarnings` or `warnings.simplefilter` with keyword arguments.

Eg, when `warnings.filterwarnings(action='ignore')` is used instead
of `warnings.filterwarnings('ignore')`, the keyword 'action'
is not accounted for in the test and the test itself runs into a list out of index
error. This should be fixed in a separate PR.